Add files via upload
This search will be in case insensitive. It will return a String of contacts holding the 
chosen last name that the user inputs. It will tell the user the number of contacts found. 
It will be in ascending order. This method will return all contacts that match the last name 
 inputed by the user.